requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.4.0
attrs==23.2.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
discord-ext-voice-recv==0.4.2a145
discord.py==2.4.0
distro==1.9.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.0
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface_hub==0.23.4
idna==3.7
jiter==0.4.2
multidict==6.0.5
mutagen==1.47.0
numpy==2.0.0
oauth2client==4.1.3
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
PyNaCl==1.5.0
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
websockets==12.0
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.7.2